feat: add `load_parameters` as a library function
