+++ manta-pay/Cargo.toml
@@ -49,6 +49,9 @@ arkworks = [
     "ark-std",
 ]
 
+# Enable Download Parameters
+download = ["manta-parameters/download"]
+
 # Enable Groth16 ZKP System
 groth16 = ["ark-groth16", "ark-snark", "arkworks"]
 
@@ -86,7 +89,7 @@ simulation = [
 std = ["manta-accounting/std", "manta-util/std"]
 
 # Testing Frameworks
-test = ["anyhow", "manta-accounting/test", "manta-crypto/test", "tempfile"]
+test = ["manta-accounting/test", "manta-crypto/test", "tempfile"]
 
 # Wallet
 wallet = ["bip32", "manta-crypto/getrandom", "std"]
@@ -104,7 +107,6 @@ websocket = [
 
 [dependencies]
 aes-gcm = { version = "0.9.4", default-features = false, features = ["aes", "alloc"] }
-anyhow = { version = "1.0.57", optional = true, default-features = false }
 ark-bls12-381 = { version = "0.3.0", optional = true, default-features = false, features = ["curve"] }
 ark-ec = { version = "0.3.0", optional = true, default-features = false }
 ark-ed-on-bls12-381 = { version = "0.3.0", optional = true, default-features = false, features = ["r1cs"] }
@@ -124,6 +126,7 @@ futures = { version = "0.3.21", optional = true, default-features = false }
 indexmap = { version = "1.8.2", optional = true, default-features = false }
 manta-accounting = { path = "../manta-accounting", default-features = false }
 manta-crypto = { path = "../manta-crypto", default-features = false }
+manta-parameters = { path = "../manta-parameters", default-features = false }
 manta-util = { path = "../manta-util", default-features = false }
 parking_lot = { version = "0.12.1", optional = true, default-features = false }
 rand_chacha = { version = "0.3.1", default-features = false }
@@ -141,5 +144,4 @@ ws_stream_wasm = { version = "0.7.3", optional = true, default-features = false 
 
 [dev-dependencies]
 manta-crypto = { path = "../manta-crypto", default-features = false, features = ["getrandom"] }
-manta-parameters = { path = "../manta-parameters", default-features = false, features = ["download"] }
-manta-pay = { path = ".", default-features = false, features = ["groth16", "scale", "scale-std", "std", "test"] }
+manta-pay = { path = ".", default-features = false, features = ["download", "groth16", "scale", "scale-std", "std", "test"] }
