+++ manta-pay/src/parameters.rs
@@ -17,11 +17,14 @@
 //! Generate Parameters and Proving/Verifying Contexts
 
 use crate::config::{
-    FullParameters, Mint, MultiProvingContext, MultiVerifyingContext, Parameters, PrivateTransfer,
-    ProofSystemError, Reclaim, UtxoAccumulatorModel,
+    FullParameters, Mint, MultiProvingContext, MultiVerifyingContext, NoteEncryptionScheme,
+    Parameters, PrivateTransfer, ProofSystemError, ProvingContext, Reclaim, UtxoAccumulatorModel,
+    UtxoCommitmentScheme, VerifyingContext, VoidNumberCommitmentScheme,
 };
 use manta_crypto::rand::{Rand, SeedableRng};
+use manta_util::codec::{Decode, IoReader};
 use rand_chacha::ChaCha20Rng;
+use std::{fs::File, path::Path};
 
 /// Parameter Generation Seed
 ///
@@ -89,4 +92,89 @@ pub fn generate() -> Result<
     ProofSystemError,
 > {
     generate_from_seed(SEED)
+}
+
+/// Loads parameters from the `manta-parameters`, using `directory` as a temporary directory to store files.
+#[cfg(feature = "download")]
+#[cfg_attr(doc_cfg, doc(cfg(feature = "download")))]
+#[inline]
+pub fn load_parameters(
+    directory: &Path,
+) -> Result<
+    (
+        MultiProvingContext,
+        MultiVerifyingContext,
+        Parameters,
+        UtxoAccumulatorModel,
+    ),
+    ProofSystemError,
+> {
+    let mint_path = directory.join("mint.dat");
+    manta_parameters::pay::testnet::proving::Mint::download(&mint_path)
+        .expect("Unable to download MINT proving context.");
+    let private_transfer_path = directory.join("private-transfer.dat");
+    manta_parameters::pay::testnet::proving::PrivateTransfer::download(&private_transfer_path)
+        .expect("Unable to download PRIVATE_TRANSFER proving context.");
+    let reclaim_path = directory.join("reclaim.dat");
+    manta_parameters::pay::testnet::proving::Reclaim::download(&reclaim_path)
+        .expect("Unable to download RECLAIM proving context.");
+    let proving_context = MultiProvingContext {
+        mint: ProvingContext::decode(IoReader(
+            File::open(mint_path).expect("Unable to read MINT proving context from disk."),
+        ))
+        .expect("Unable to decode MINT proving context."),
+        private_transfer: ProvingContext::decode(IoReader(
+            File::open(private_transfer_path)
+                .expect("Unable to read PRIVATE_TRANSFER proving context from disk."),
+        ))
+        .expect("Unable to decode PRIVATE_TRANSFER proving context."),
+        reclaim: ProvingContext::decode(IoReader(
+            File::open(reclaim_path).expect("Unable to read RECLAIM provin context from disk."),
+        ))
+        .expect("Unable to decode RECLAIM proving context."),
+    };
+    let verifying_context = MultiVerifyingContext {
+        mint: VerifyingContext::decode(
+            manta_parameters::pay::testnet::verifying::Mint::get()
+                .expect("Checksum did not match."),
+        )
+        .expect("Unable to decode MINT verifying context."),
+        private_transfer: VerifyingContext::decode(
+            manta_parameters::pay::testnet::verifying::PrivateTransfer::get()
+                .expect("Checksum did not match."),
+        )
+        .expect("Unable to decode PRIVATE_TRANSFER verifying context."),
+        reclaim: VerifyingContext::decode(
+            manta_parameters::pay::testnet::verifying::Reclaim::get()
+                .expect("Checksum did not match."),
+        )
+        .expect("Unable to decode RECLAIM verifying context."),
+    };
+    let parameters = Parameters {
+        note_encryption_scheme: NoteEncryptionScheme::decode(
+            manta_parameters::pay::testnet::parameters::NoteEncryptionScheme::get()
+                .expect("Checksum did not match."),
+        )
+        .expect("Unable to decode NOTE_ENCRYPTION_SCHEME parameters."),
+        utxo_commitment: UtxoCommitmentScheme::decode(
+            manta_parameters::pay::testnet::parameters::UtxoCommitmentScheme::get()
+                .expect("Checksum did not match."),
+        )
+        .expect("Unable to decode UTXO_COMMITMENT_SCHEME parameters."),
+        void_number_commitment: VoidNumberCommitmentScheme::decode(
+            manta_parameters::pay::testnet::parameters::VoidNumberCommitmentScheme::get()
+                .expect("Checksum did not match."),
+        )
+        .expect("Unable to decode VOID_NUMBER_COMMITMENT_SCHEME parameters."),
+    };
+    Ok((
+        proving_context,
+        verifying_context,
+        parameters,
+        UtxoAccumulatorModel::decode(
+            manta_parameters::pay::testnet::parameters::UtxoAccumulatorModel::get()
+                .expect("Checksum did not match."),
+        )
+        .expect("Unable to decode UTXO_ACCUMULATOR_MODEL."),
+    ))
 }
