+++ manta-pay/src/test/compatibility.rs
@@ -18,91 +18,10 @@
 //! Checks if the current circuit implementation is compatible with precomputed parameters.
 
 use crate::{
-    config::{
-        MultiProvingContext, MultiVerifyingContext, NoteEncryptionScheme, Parameters,
-        ProvingContext, UtxoAccumulatorModel, UtxoCommitmentScheme, VerifyingContext,
-        VoidNumberCommitmentScheme,
-    },
-    test::payment::{prove_mint, prove_private_transfer, prove_reclaim},
+    test::payment::{prove_mint, prove_private_transfer, prove_reclaim}, parameters::load_parameters,
 };
-use anyhow::Result;
 use manta_accounting::transfer::test::assert_valid_proof;
 use manta_crypto::rand::{OsRng, Rand};
-use manta_util::codec::{Decode, IoReader};
-use std::{fs::File, path::Path};
-
-/// Loads parameters from the `manta-parameters`, using `directory` as a temporary directory to store files.
-#[inline]
-fn load_parameters(
-    directory: &Path,
-) -> Result<(
-    MultiProvingContext,
-    MultiVerifyingContext,
-    Parameters,
-    UtxoAccumulatorModel,
-)> {
-    println!("Loading parameters...");
-    let mint_path = directory.join("mint.dat");
-    manta_parameters::pay::testnet::proving::Mint::download(&mint_path)?;
-    let private_transfer_path = directory.join("private-transfer.dat");
-    manta_parameters::pay::testnet::proving::PrivateTransfer::download(&private_transfer_path)?;
-    let reclaim_path = directory.join("reclaim.dat");
-    manta_parameters::pay::testnet::proving::Reclaim::download(&reclaim_path)?;
-    println!("mint_path: {:?}", mint_path);
-    let proving_context = MultiProvingContext {
-        mint: ProvingContext::decode(IoReader(File::open(mint_path)?))
-            .expect("Unable to decode MINT proving context."),
-        private_transfer: ProvingContext::decode(IoReader(File::open(private_transfer_path)?))
-            .expect("Unable to decode PRIVATE_TRANSFER proving context."),
-        reclaim: ProvingContext::decode(IoReader(File::open(reclaim_path)?))
-            .expect("Unable to decode RECLAIM proving context."),
-    };
-    let verifying_context = MultiVerifyingContext {
-        mint: VerifyingContext::decode(
-            manta_parameters::pay::testnet::verifying::Mint::get()
-                .expect("Checksum did not match."),
-        )
-        .expect("Unable to decode MINT verifying context."),
-        private_transfer: VerifyingContext::decode(
-            manta_parameters::pay::testnet::verifying::PrivateTransfer::get()
-                .expect("Checksum did not match."),
-        )
-        .expect("Unable to decode PRIVATE_TRANSFER verifying context."),
-        reclaim: VerifyingContext::decode(
-            manta_parameters::pay::testnet::verifying::Reclaim::get()
-                .expect("Checksum did not match."),
-        )
-        .expect("Unable to decode RECLAIM verifying context."),
-    };
-    let parameters = Parameters {
-        note_encryption_scheme: NoteEncryptionScheme::decode(
-            manta_parameters::pay::testnet::parameters::NoteEncryptionScheme::get()
-                .expect("Checksum did not match."),
-        )
-        .expect("Unable to decode NOTE_ENCRYPTION_SCHEME parameters."),
-        utxo_commitment: UtxoCommitmentScheme::decode(
-            manta_parameters::pay::testnet::parameters::UtxoCommitmentScheme::get()
-                .expect("Checksum did not match."),
-        )
-        .expect("Unable to decode UTXO_COMMITMENT_SCHEME parameters."),
-        void_number_commitment: VoidNumberCommitmentScheme::decode(
-            manta_parameters::pay::testnet::parameters::VoidNumberCommitmentScheme::get()
-                .expect("Checksum did not match."),
-        )
-        .expect("Unable to decode VOID_NUMBER_COMMITMENT_SCHEME parameters."),
-    };
-    println!("Loading parameters Done.");
-    Ok((
-        proving_context,
-        verifying_context,
-        parameters,
-        UtxoAccumulatorModel::decode(
-            manta_parameters::pay::testnet::parameters::UtxoAccumulatorModel::get()
-                .expect("Checksum did not match."),
-        )
-        .expect("Unable to decode UTXO_ACCUMULATOR_MODEL."),
-    ))
-}
 
 /// Tests that the circuit is compatible with the current known parameters in `manta-parameters`.
 #[test]
